requirements.txt
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
django-redis==4.12.1
djangorestframework==3.12.2
idna==2.10
pytz==2020.4
redis==3.5.3
requests==2.25.0
urllib3==1.26.2

